consolidate render stuff in systems/render
