+++ src/main/java/com/marginallyclever/robotoverlord/systems/render/Compass3D.java
@@ -1,4 +1,4 @@
-package com.marginallyclever.robotoverlord;
+package com.marginallyclever.robotoverlord.systems.render;
 
 import com.jogamp.opengl.GL2;
 import com.marginallyclever.convenience.MatrixHelper;
