+++ src/main/java/com/marginallyclever/robotoverlord/systems/render/SkyBox.java
@@ -1,13 +1,11 @@
-package com.marginallyclever.robotoverlord;
+package com.marginallyclever.robotoverlord.systems.render;
 
 import com.jogamp.opengl.GL2;
 import com.marginallyclever.convenience.MatrixHelper;
 import com.marginallyclever.convenience.OpenGLHelper;
 import com.marginallyclever.robotoverlord.components.CameraComponent;
 import com.marginallyclever.robotoverlord.components.PoseComponent;
-import com.marginallyclever.robotoverlord.components.shapes.Box;
 import com.marginallyclever.robotoverlord.parameters.TextureParameter;
-import com.marginallyclever.robotoverlord.systems.render.mesh.Mesh;
 
 import javax.vecmath.Matrix4d;
 import javax.vecmath.Vector3d;
