+++ src/main/java/com/marginallyclever/robotoverlord/tools/move/MoveCameraTool.java
@@ -2,7 +2,7 @@ package com.marginallyclever.robotoverlord.tools.move;
 
 import com.jogamp.opengl.GL2;
 import com.marginallyclever.robotoverlord.entity.Entity;
-import com.marginallyclever.robotoverlord.Viewport;
+import com.marginallyclever.robotoverlord.systems.render.Viewport;
 import com.marginallyclever.robotoverlord.components.CameraComponent;
 import com.marginallyclever.robotoverlord.parameters.DoubleParameter;
 import com.marginallyclever.robotoverlord.tools.EditorTool;
