+++ src/main/java/com/marginallyclever/robotoverlord/tools/move/TranslateEntityMultiTool.java
@@ -3,7 +3,7 @@ package com.marginallyclever.robotoverlord.tools.move;
 import com.jogamp.opengl.GL2;
 import com.marginallyclever.convenience.MatrixHelper;
 import com.marginallyclever.robotoverlord.entity.Entity;
-import com.marginallyclever.robotoverlord.Viewport;
+import com.marginallyclever.robotoverlord.systems.render.Viewport;
 import com.marginallyclever.robotoverlord.components.PoseComponent;
 import com.marginallyclever.robotoverlord.tools.EditorTool;
