+++ src/test/java/com/marginallyclever/robotoverlord/ViewportTest.java
@@ -1,5 +1,6 @@
 package com.marginallyclever.robotoverlord;
 
+import com.marginallyclever.robotoverlord.systems.render.Viewport;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
